+++ src/events/mod.rs
@@ -1,6 +1,0 @@
-// Copyright (C) 2019-2021 The apca Developers
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-mod stream;
-
-pub use stream::connect;